Refactor the resource cross-language test to a cross-language test in CI.
