+++ python/flink_agents/e2e_tests/e2e_tests_integration/chat_model_integration_agent.py
