+++ python/flink_agents/e2e_tests/e2e_tests_resource_cross_language/chat_model_cross_language_test.py
@@ -16,6 +16,7 @@
 # limitations under the License.
 #################################################################################
 import os
+import sysconfig
 from pathlib import Path
 
 import pytest
@@ -43,6 +44,8 @@ OLLAMA_MODEL = os.environ.get("OLLAMA_CHAT_MODEL", "qwen3:1.7b")
 os.environ["OLLAMA_CHAT_MODEL"] = OLLAMA_MODEL
 
 client = pull_model(OLLAMA_MODEL)
+
+os.environ["PYTHONPATH"] = sysconfig.get_paths()["purelib"]
 
 @pytest.mark.skipif(client is None, reason="Ollama client is not available or test model is missing.")
 def test_java_chat_model_integration(tmp_path: Path) -> None:  # noqa: D103
